@@ -1,14 +1,14 @@
 plugins {
     id 'java'
     id 'eclipse'
-    id 'org.gradle.java' 
+    id 'org.gradle.java'
     id 'org.springframework.boot' version '3.2.0'
     id 'io.spring.dependency-management' version '1.1.0'
 }
 
 group = 'com.mytech.apartment.portal'
 version = '0.0.1-SNAPSHOT'
-sourceCompatibility = '17'
+sourceCompatibility = '20'
 
 repositories {
     mavenCentral()
@@ -19,7 +19,7 @@ compileJava {
     options.compilerArgs += ['-parameters']
 }
 
-// ĐẨY annotationProcessor JAR vào Eclipse Factory Path
+// Đẩy annotationProcessor JAR vào Eclipse Factory Path
 eclipse {
     classpath {
         plusConfigurations += [ configurations.annotationProcessor ]
@@ -39,7 +39,7 @@ dependencies {
     compileOnly 'org.projectlombok:lombok'
     annotationProcessor 'org.projectlombok:lombok'
 
-    // Cập nhật MapStruct lên phiên bản mới hơn
+    // MapStruct phiên bản mới nhất tương thích với Java 20
     implementation 'org.mapstruct:mapstruct:1.6.2'
     annotationProcessor 'org.mapstruct:mapstruct-processor:1.6.2'
 
