@@ -2,6 +2,7 @@ package com.mytech.apartment.portal.mappers;
 
 import com.mytech.apartment.portal.dtos.InvoiceDto;
 import com.mytech.apartment.portal.models.Invoice;
+import com.mytech.apartment.portal.models.enums.InvoiceStatus;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
@@ -24,7 +25,7 @@ public class InvoiceMapper {
         dto.setIssueDate(entity.getIssueDate());
         dto.setDueDate(entity.getDueDate());
         dto.setTotalAmount(entity.getTotalAmount());
-        dto.setStatus(entity.getStatus());
+        dto.setStatus(entity.getStatus() != null ? entity.getStatus().name() : null);
         dto.setCreatedAt(entity.getCreatedAt());
         dto.setUpdatedAt(entity.getUpdatedAt());
         if (entity.getItems() != null) {
@@ -33,5 +34,27 @@ public class InvoiceMapper {
                     .collect(Collectors.toSet()));
         }
         return dto;
+    }
+
+    public Invoice fromDto(InvoiceDto dto) {
+        if (dto == null) {
+            return null;
+        }
+        Invoice entity = new Invoice();
+        entity.setId(dto.getId());
+        entity.setApartmentId(dto.getApartmentId());
+        entity.setBillingPeriod(dto.getBillingPeriod());
+        entity.setIssueDate(dto.getIssueDate());
+        entity.setDueDate(dto.getDueDate());
+        entity.setTotalAmount(dto.getTotalAmount());
+        entity.setStatus(dto.getStatus() != null ? InvoiceStatus.valueOf(dto.getStatus()) : null);
+        entity.setCreatedAt(dto.getCreatedAt());
+        entity.setUpdatedAt(dto.getUpdatedAt());
+        if (dto.getItems() != null) {
+            entity.setItems(dto.getItems().stream()
+                    .map(invoiceItemMapper::toEntity)
+                    .collect(Collectors.toSet()));
+        }
+        return entity;
     }
 } 
