@@ -1,8 +1,9 @@
 package com.mytech.apartment.portal.mappers;
 
+import org.springframework.stereotype.Component;
+
 import com.mytech.apartment.portal.dtos.PaymentDto;
 import com.mytech.apartment.portal.models.Payment;
-import org.springframework.stereotype.Component;
 
 @Component
 public class PaymentMapper {
@@ -19,8 +20,8 @@ public class PaymentMapper {
         dto.setPaidByUserId(entity.getPaidByUserId());
         dto.setPaymentDate(entity.getPaymentDate());
         dto.setAmount(entity.getAmount());
-        dto.setMethod(entity.getMethod());
-        dto.setStatus(entity.getStatus());
+        dto.setMethod(entity.getMethod() != null ? entity.getMethod().name() : null);
+        dto.setStatus(entity.getStatus() != null ? entity.getStatus().name() : null);
         dto.setReferenceCode(entity.getReferenceCode());
         return dto;
     }
