@@ -3,18 +3,14 @@ package com.mytech.apartment.portal.mappers;
 import com.mytech.apartment.portal.dtos.UserCreateRequest;
 import com.mytech.apartment.portal.dtos.UserDto;
 import com.mytech.apartment.portal.dtos.UserUpdateRequest;
-import com.mytech.apartment.portal.models.Role;
 import com.mytech.apartment.portal.models.User;
-import org.springframework.beans.factory.annotation.Autowired;
+import com.mytech.apartment.portal.models.enums.UserStatus;
 import org.springframework.stereotype.Component;
 
 import java.util.stream.Collectors;
 
 @Component
 public class UserMapper {
-
-    @Autowired
-    private RoleMapper roleMapper;
 
     public UserDto toDto(User user) {
         if (user == null) {
@@ -24,14 +20,11 @@ public class UserMapper {
         dto.setId(user.getId());
         dto.setUsername(user.getUsername());
         dto.setPhoneNumber(user.getPhoneNumber());
-        dto.setStatus(user.getStatus());
+        dto.setStatus(user.getStatus() != null ? user.getStatus().name() : null);
         dto.setCreatedAt(user.getCreatedAt());
         dto.setUpdatedAt(user.getUpdatedAt());
         dto.setLockReason(user.getLockReason());
         dto.setEmail(user.getEmail());
-        if (user.getRoles() != null) {
-            dto.setRoles(user.getRoles().stream().map(Role::getName).collect(Collectors.toSet()));
-        }
         return dto;
     }
 
@@ -43,7 +36,7 @@ public class UserMapper {
         user.setUsername(request.getUsername());
         user.setPasswordHash(request.getPassword());
         user.setPhoneNumber(request.getPhoneNumber());
-        user.setStatus("ACTIVE"); // Default status
+        user.setStatus(UserStatus.ACTIVE); // Default status
         return user;
     }
 
@@ -55,7 +48,7 @@ public class UserMapper {
             user.setPhoneNumber(request.getPhoneNumber());
         }
         if (request.getStatus() != null) {
-            user.setStatus(request.getStatus());
+            user.setStatus(UserStatus.valueOf(request.getStatus()));
         }
     }
 } 
