@@ -1,6 +1,7 @@
 package com.mytech.apartment.portal.repositories;
 
 import com.mytech.apartment.portal.models.Payment;
+import com.mytech.apartment.portal.models.enums.PaymentStatus;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -11,5 +12,5 @@ import java.util.Optional;
 public interface PaymentRepository extends JpaRepository<Payment, Long> {
     Optional<Payment> findByReferenceCode(String referenceCode);
     List<Payment> findByInvoiceId(Long invoiceId);
-    List<Payment> findByStatus(String status);
+    List<Payment> findByStatus(PaymentStatus status);
 }
