@@ -4,9 +4,7 @@ import com.mytech.apartment.portal.models.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.Optional;
-
 @Repository
-public interface RoleRepository extends JpaRepository<Role, Integer> {
-    Optional<Role> findByName(String name);
+public interface RoleRepository extends JpaRepository<Role, Long> {
+    Role findByName(String name);
 } 
