@@ -1,6 +1,8 @@
 package com.mytech.apartment.portal.repositories;
 
 import com.mytech.apartment.portal.models.Payment;
+import com.mytech.apartment.portal.models.enums.PaymentMethod;
+import com.mytech.apartment.portal.models.enums.PaymentStatus;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -19,13 +21,13 @@ public class PaymentRepositoryTest {
                 .invoice(null)
                 .paidByUserId(1L)
                 .amount(500000.0)
-                .method("CASH")
-                .status("SUCCESS")
+                .method(PaymentMethod.CASH)
+                .status(PaymentStatus.SUCCESS)
                 .build();
         Payment saved = paymentRepository.save(payment);
         Optional<Payment> found = paymentRepository.findById(saved.getId());
         Assertions.assertTrue(found.isPresent());
-        Assertions.assertEquals("SUCCESS", found.get().getStatus());
+        Assertions.assertEquals(PaymentStatus.SUCCESS, found.get().getStatus());
     }
 
     @Test
@@ -34,8 +36,8 @@ public class PaymentRepositoryTest {
                 .invoice(null)
                 .paidByUserId(2L)
                 .amount(700000.0)
-                .method("BANK")
-                .status("SUCCESS")
+                .method(PaymentMethod.BANK)
+                .status(PaymentStatus.SUCCESS)
                 .build();
         Payment saved = paymentRepository.save(payment);
         paymentRepository.deleteById(saved.getId());
